aoc_2025_12::solve

Function find_one_fulfillment

pub fn find_one_fulfillment<const S: usize>(
    requirement: &Requirement<S>,
    placements: &[Placement<'_, S>],
) -> Result<Option<Vec<usize>>>

Using the given pre-computed placements of shapes, perform a Dancing Links to determine if the requirement can be fulfilled.

This is performed by

DFS through the placements without duplicates, keeping track of the

current path of placements,

a single mutable StateStorage representing the current fulfillment state

a nested Vec to track eliminated placements at each depth level;

everytime we try to advance the path, we

update the current state storage by applying the placement at the current path’s last index

check if the instance count portion of the StateStorage are all 1